move

+++ app/Models/Staff.php
@@ -12,7 +12,7 @@ class Staff extends Model
     protected static function booted()
     {
         static::deleting(function ($staff) {
-            if ($staff->photo && ($staff->photo !== '/default/staff-male.svg' || $staff->photo !== '/default/staff-female.svg')) {
+            if ($staff->photo && ($staff->photo !== 'default/staff-male.svg' && $staff->photo !== 'default/staff-female.svg')) {
                 Storage::disk('public')->delete($staff->photo);
             }
         });
@@ -20,7 +20,7 @@ class Staff extends Model
         static::updating(function ($staff) {
             if ($staff->isDirty('photo')) {
                 $oldPhoto = $staff->getOriginal('photo');
-                if ($oldPhoto && ($oldPhoto !== '/default/staff-male.svg' || $oldPhoto !== '/default/staff-female.svg')) {
+                if ($oldPhoto && ($oldPhoto !== 'default/staff-male.svg' && $oldPhoto !== 'default/staff-female.svg')) {
                     Storage::disk('public')->delete($oldPhoto);
                 }
             }
